+++ core-ng-search-test/src/main/java/core/framework/search/module/TestSearchConfig.java
@@ -2,12 +2,9 @@ package core.framework.search.module;
 
 import core.framework.internal.module.ModuleContext;
 import core.framework.internal.module.ShutdownHook;
-import core.framework.search.impl.ESLoggerConfigFactory;
 import core.framework.search.impl.LocalElasticSearch;
-import org.apache.http.HttpHost;
-import org.apache.logging.log4j.core.config.ConfigurationFactory;
-import org.elasticsearch.common.logging.internal.LoggerFactoryImpl;
-import org.elasticsearch.logging.internal.spi.LoggerFactory;
+import org.apache.hc.core5.http.HttpHost;
+import org.elasticsearch.common.logging.LogConfigurator;
 
 import java.util.concurrent.locks.ReentrantLock;
 
@@ -31,7 +28,6 @@ public class TestSearchConfig extends SearchConfig {
     @Override
     protected void initialize(ModuleContext context, String name) {
         super.initialize(context, name);
-        configureLogger();
         startLocalElasticSearch(context);
     }
 
@@ -40,6 +36,8 @@ public class TestSearchConfig extends SearchConfig {
         try {
             // in test env, config is initialized in order and within same thread, so no threading issue
             if (localESHost == null) {
+                LogConfigurator.configureESLogging();
+
                 var server = new LocalElasticSearch();
                 localESHost = server.start();
                 context.shutdownHook.add(ShutdownHook.STAGE_6, timeout -> server.close());
@@ -56,27 +54,5 @@ public class TestSearchConfig extends SearchConfig {
 
     @Override
     public void auth(String apiKeyId, String apiKeySecret) {
-
-    }
-
-    // ES uses log4j2 core api directly, cannot use log4j-to-slf4j to bridge, refer to following exception
-    // here is to bridge to core-ng logger
-    /*
-      java.lang.ClassCastException: class org.apache.logging.slf4j.SLF4JLoggerContext cannot be cast to class org.apache.logging.log4j.core.LoggerContext (org.apache.logging.slf4j.SLF4JLoggerContext and org.apache.logging.log4j.core.LoggerContext are in unnamed module of loader 'app')
-        at org.apache.logging.log4j.core.LoggerContext.getContext(LoggerContext.java:231)
-        at org.apache.logging.log4j.core.config.Configurator.setLevel(Configurator.java:366)
-        at org.elasticsearch.common.logging.Loggers.setLevel(Loggers.java:114)
-        at org.elasticsearch.index.SearchSlowLog.<init>(SearchSlowLog.java:111)
-        at org.elasticsearch.index.IndexModule.<init>(IndexModule.java:176)
-    * */
-    void configureLogger() {
-        // ES starts to migrate log4j api to its own logger api, once important info migrated (like node info), we will remove log4j adapter
-        if (LoggerFactory.provider() == null) LoggerFactory.setInstance(new LoggerFactoryImpl());
-
-        if (System.getProperty(ConfigurationFactory.CONFIGURATION_FACTORY_PROPERTY) != null) return;
-        System.setProperty(ConfigurationFactory.CONFIGURATION_FACTORY_PROPERTY, ESLoggerConfigFactory.class.getName());
-        // refer to org.apache.logging.log4j.core.jmx.Server.PROPERTY_DISABLE_JMX, disable to reduce overhead
-        System.setProperty("log4j2.disable.jmx", "true");
-        ESLoggerConfigFactory.configureLogger();
     }
 }