+++ core-ng-search-test/src/main/java/org/elasticsearch/node/LocalNode.java
@@ -6,10 +6,12 @@ import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.mapper.extras.MapperExtrasPlugin;
 import org.elasticsearch.painless.PainlessPlugin;
 import org.elasticsearch.plugins.LocalPluginsService;
+import org.elasticsearch.plugins.PluginsLoader;
 import org.elasticsearch.plugins.PluginsService;
 import org.elasticsearch.reindex.ReindexPlugin;
 import org.elasticsearch.transport.netty4.Netty4Plugin;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -17,10 +19,10 @@ import java.util.List;
  */
 public class LocalNode extends Node {
     public LocalNode(Settings settings) {
-        super(NodeConstruction.prepareConstruction(new Environment(settings, null), new NodeServiceProvider() {
+        super(NodeConstruction.prepareConstruction(new Environment(settings, null), PluginsLoader.createPluginsLoader(Collections.emptySet(), Collections.emptySet(), Collections.emptyMap(), false), new NodeServiceProvider() {
             @Override
-            PluginsService newPluginService(Environment environment, Settings settings) {
-                return new LocalPluginsService(settings,
+            PluginsService newPluginService(Environment initialEnvironment, PluginsLoader pluginsLoader) {
+                return new LocalPluginsService(settings, pluginsLoader,
                     List.of(Netty4Plugin.class,           // for http transport
                         MapperExtrasPlugin.class,         // for scaled_float type
                         PainlessPlugin.class,