+++ core-ng-search-test/src/main/java/org/elasticsearch/plugins/LocalPluginsService.java
@@ -2,12 +2,10 @@ package org.elasticsearch.plugins;
 
 import org.elasticsearch.Version;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.jdk.ModuleQualifiedExportsService;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 
 /**
  * @author neo
@@ -18,8 +16,8 @@ public class LocalPluginsService extends PluginsService {
     // refer to https://github.com/elastic/elasticsearch/blob/main/test/framework/src/main/java/org/elasticsearch/plugins/MockPluginsService.java
     // refer to https://github.com/elastic/elasticsearch/blob/main/test/framework/src/main/java/org/elasticsearch/node/MockNode.java
     // refer to https://github.com/elastic/elasticsearch/blob/main/test/framework/src/main/java/org/elasticsearch/test/ESSingleNodeTestCase.java
-    public LocalPluginsService(Settings settings, Collection<Class<? extends Plugin>> pluginClasses) {
-        super(settings, null, null, null);
+    public LocalPluginsService(Settings settings, PluginsLoader loader, Collection<Class<? extends Plugin>> pluginClasses) {
+        super(settings, null, loader);
         plugins = new ArrayList<>(pluginClasses.size());
         for (Class<? extends Plugin> pluginClass : pluginClasses) {
             Plugin plugin = loadPlugin(pluginClass, settings, null);
@@ -28,7 +26,7 @@ public class LocalPluginsService extends PluginsService {
                 "classpath plugin",
                 "NA",
                 Version.CURRENT.toString(),
-                "21",
+                "24",
                 pluginClass.getName(),
                 null,
                 List.of(),
@@ -36,7 +34,7 @@ public class LocalPluginsService extends PluginsService {
                 false,
                 false,
                 false);
-            plugins.add(new LoadedPlugin(descriptor, plugin, Thread.currentThread().getContextClassLoader(), ModuleLayer.boot()));
+            plugins.add(new LoadedPlugin(descriptor, plugin, Thread.currentThread().getContextClassLoader()));
         }
         loadExtensions(plugins);
     }
@@ -44,9 +42,5 @@ public class LocalPluginsService extends PluginsService {
     @Override
     protected final List<LoadedPlugin> plugins() {
         return this.plugins;
-    }
-
-    @Override
-    protected void addServerExportsService(Map<String, List<ModuleQualifiedExportsService>> qualifiedExports) {
     }
 }