+++ core-ng-search-test/src/test/java/core/framework/search/module/TestSearchConfigTest.java
@@ -3,8 +3,6 @@ package core.framework.search.module;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 /**
  * @author neo
  */
@@ -17,9 +15,7 @@ class TestSearchConfigTest {
     }
 
     @Test
-    void configureLogger() {
-        config.configureLogger();
-
-        assertThat(System.getProperty("log4j.configurationFactory")).isNotNull();
+    void auth() {
+        config.auth("id", "secret");
     }
 }